main/ai-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sentence-transformers==2.6.1
transformers>=4.38.0
chromadb==0.4.24
groq==0.5.0
torch>=2.0.0
psycopg2-binary==2.9.9
redis==5.0.3
python-dotenv==1.0.1
numpy==1.26.4
httpx==0.27.0
pydantic==2.6.4
